txHandler: move streamverifer out of txn go (#5039)
